+++ libpok/include/core/thread.h
@@ -36,8 +36,8 @@ typedef struct {
   uint32_t stack_size;
   uint32_t state;
   bool_t user_add;
-  uint8_t finish_num;
-  uint8_t miss_num;
+  uint32_t finish_num;
+  uint32_t miss_num;
   uint32_t total_num;
   uint64_t soft_deadline;
 } pok_thread_attr_t;
